make floorid int again and fixed int to float conversions

+++ thesisbackend/src/main/java/com/thesis/backend/model/dto/CompartmentDto.java
@@ -16,5 +16,5 @@ public class CompartmentDto {
     private float yKonva;
     private float widthKonva;
     private float heightKonva;
-    private float floorId;
+    private int floorId;
 }
